add debugging & 88
