+++ .vscode/launch.json
@@ -1,0 +1,20 @@
+{
+  // Use IntelliSense to learn about possible attributes.
+  // Hover to view descriptions of existing attributes.
+  // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+  "version": "0.2.0",
+  "configurations": [
+    {
+      "command": "npm run build && npm run start",
+      "name": "debugging",
+      "request": "launch",
+      "type": "node-terminal",
+      "console": "integratedTerminal",
+      "cwd": "${workspaceFolder}",
+      "resolveSourceMapLocations": [
+        "${workspaceFolder}/build/**/*.js",
+        "!**/node_modules/**"
+      ]
+    }
+  ]
+}
