+++ src/index.ts
@@ -1,0 +1,3 @@
+import {merge} from './88-merge-sorted-array/merge-sorted-array.js'
+
+merge([1, 2, 3, 0, 0, 0], 3, [2, 5, 6], 3)
